pyproject.toml
[tool.poetry]
name = "arubacx-pdf-extractor"
version = "0.1.0"
description = ""
authors = ["Colin Heurteaux <dev@heurteaux.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
openai = "^1.54.1"
python-dotenv = "^1.0.1"
alive-progress = "^3.2.0"
halo = "^0.0.31"
pypdf = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
